Refactoring: naming, dates, duration
