+++ src/view/trip-filters.js
