+++ src/view/trip-menu.js
