+++ src/view/trip-point-list.js
@@ -1,4 +1,4 @@
-export const createTripEventsList = () => {
+export const createPointsList = () => {
   return `<ul class="trip-events__list">
  </ul>`;
 };
