+++ src/view/trip-sorting.js
@@ -1,4 +1,4 @@
-export const createTripSort = () => {
+export const createSort = () => {
   return `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
   <div class="trip-sort__item  trip-sort__item--day">
     <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day" checked>
